finished! :D

+++ controllers/courseController.js
@@ -24,6 +24,10 @@ module.exports.viewProfile= async function(req,res){
 
 //render add form
 module.exports.renderAddForm =  function(req, res){
+    if (!req.user.can('add course')) {
+        res.redirect('/');
+        return
+    }
     const course =  {
         name: '',
         department: departments[0],
@@ -35,6 +39,10 @@ module.exports.renderAddForm =  function(req, res){
 
 //add
 module.exports.addCourse = async function(req, res){
+    if (!req.user.can('add course')) {
+        res.redirect('/');
+        return
+    }
     const course = await Course.create({
         name: req.body.name,
         department: req.body.department,
@@ -46,12 +54,20 @@ module.exports.addCourse = async function(req, res){
 
 //render edit form
 module.exports.renderEditForm = async function(req, res){
+    if (!req.user.can('edit course')) {
+        res.redirect('/');
+        return
+    }
     const course = await Course.findByPk(req.params.id);
     res.render('course/edit', {course, departments});
 }
 
 //update
 module.exports.updateCourse = async function(req, res){
+    if (!req.user.can('edit course')) {
+        res.redirect('/');
+        return
+    }
     const course = await Course.update({
         name: req.body.name,
         department: req.body.department,
@@ -67,6 +83,10 @@ module.exports.updateCourse = async function(req, res){
 
 //delete
 module.exports.deleteCourse = async function(req, res){
+    if (!req.user.can('delete course')) {
+        res.redirect('/');
+        return
+    }
     await Course.destroy({
         where: {
             id:req.params.id
@@ -77,6 +97,15 @@ module.exports.deleteCourse = async function(req, res){
 
 //Add student to a course
 module.exports.enrollStudent = async function(req, res){
+    const isAdmin = req.user.can('enroll students');
+    const profileBelongsToUser = req.user.can('enroll self') && req.user.matchesStudentId(req.body.student);
+
+    // if not staff or student, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
+
     await StudentCourses.create({
         student_id: req.body.student,
         course_id: req.params.courseId
@@ -86,6 +115,15 @@ module.exports.enrollStudent = async function(req, res){
 
 //remove a student from a course
 module.exports.removeStudent = async function(req, res){
+    const isAdmin = req.user.can('drop students');
+    const profileBelongsToUser = req.user.can('drop self') && req.user.matchesStudentId(req.params.studentId);
+
+    // if not staff or student, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
+
     await StudentCourses.destroy({
         where: {
             course_id: req.params.courseId,
